requirements.txt
absl-py==1.4.0
certifi==2023.5.7
charset-normalizer==3.1.0
contourpy==1.1.0
cycler==0.11.0
fonttools==4.40.0
idna==3.4
joblib==1.3.1
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.25.0
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
pyparsing==3.1.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
scikit-learn==1.0.2
scipy==1.11.1
six==1.16.0
threadpoolctl==3.1.0
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchmetrics=0.9.3
torchvision==0.12.0+cu113
typing_extensions==4.7.0
tzdata==2023.3
urllib3==2.0.3
wfdb==3.4.1
